==> code/08-prepare-cox-summary-table.ipynb <==
suppressPackageStartupMessages(library(tidyverse))
library(lubridate)
library(glue)
# library(bupaR)
# library(processanimateR)
library(survival)
library(ggplot2)
library(ggfortify)
library(patchwork)
library(ggsurvfit)
library(survminer)
library(gtsummary)
suppressPackageStartupMessages(library(flextable))
suppressPackageStartupMessages(library(Greg))

est_ci_fn <- function(df) {
    as.character(df %>% reframe(glue::glue("{round(estimate, 2)} ({round(conf.low, 2)}--{round(conf.high,2)})")) %>% pull())
}

regression_summary <- function(grouping_number, ed_adn_type = "non_avoid") {

    basic_avoid_df <- readRDS(glue::glue("output/grouping{grouping_number}_cox_basic_non_avoid.rds"))
    adj_avoid_df <- readRDS(glue::glue("output/grouping{grouping_number}_cox_adj_non_avoid.rds"))
    basic_all_df <- readRDS(glue::glue("output/grouping{grouping_number}_cox_basic_all.rds"))
    adj_all_df <- readRDS(glue::glue("output/grouping{grouping_number}_cox_adj_all.rds"))
    summary_df <- readRDS(glue::glue("output/grouping{grouping_number}_summary_df.rds")) %>%
        count(strata)
    
    # mutate(across(telex:players, ~ format(., big.mark = ",", scientific = F)))
    un_hr_avoid = est_ci_fn(basic_avoid_df[[2]])
    adj_hr_avoid = est_ci_fn(adj_avoid_df[[2]][1,])
    un_hr_all = est_ci_fn(basic_all_df[[2]])
    adj_hr_all = est_ci_fn(adj_all_df[[2]][1,])

    #print(un_hr_avoid)
    
    result <- tibble(
        grouping = glue::glue("Grouping {grouping_number}"),
        abn = summary_df$n[summary_df$strata == "ABN"],
        iuc = summary_df$n[summary_df$strata == "IUC"],
        un_hr_avoid = un_hr_avoid,
        adj_hr_avoid = adj_hr_avoid,
        un_hr_all = un_hr_all,
        adj_hr_all = adj_hr_all
    )

    result
    
    
}

cox_summary_tbl <- dplyr::bind_rows(
    regression_summary(1),
    regression_summary(2),
    regression_summary(3)
)

cox_summary_tbl

cst <- cox_summary_tbl %>%
    rename(
        Grouping = grouping,
        `Abandoned calls (N)` = abn,
        `Triaged 111 calls (N)` = iuc,
        `Unadjusted HR (95% CI)` = un_hr_avoid,
        `Adjusted HR (95% CI)` = adj_hr_avoid,
        `Unadjusted HR (95%CI)` = un_hr_all,
        `Adjusted HR (95%CI)` = adj_hr_all
    ) %>%
    as_flextable() %>%
    add_header_row(
      values = c("", "Non-avoidable ED attendances", "All ED attendances"),
      colwidths = c(3, 2, 2)
    )

cst

saveRDS(cst, 'output/cox_regression_flex_table.rds')

cst %>%
    save_as_docx(path = glue::glue("output/cox_regression_table.docx"))
