==> code/09-prepare-summary-km-plots.ipynb <==
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(patchwork))
library(ggsurvfit)

km_1 <- readRDS('output/km_non_avoid_plot_grouping1.rds')
km_2 <- readRDS('output/km_non_avoid_plot_grouping2.rds')
km_3 <- readRDS('output/km_non_avoid_plot_grouping3.rds')

options(repr.plot.width=14, repr.plot.height=6)

final_plot <- patchwork::wrap_plots(
    wrap_elements(km_1), 
    wrap_elements(km_2),
    wrap_elements(km_3),
    ncol = 3
)
ggsave(plot = final_plot, file = glue::glue("output/km_plot_summary.pdf"), width = 14)
final_plot
